platform/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
numpy==1.26.2
playwright==1.40.0
discord.py==2.3.2
python-telegram-bot==20.7
APScheduler==3.10.4
